requirements.txt
llama-index
llama-index-llms-ollama
llama-index-embeddings-huggingface
python-dotenv